Exploded editor-state into more manageable chunk of code.
The newly introduce classes are EditorArea, PaletteArea and Brush.
More responsability was given to the Brush to make it simpler to
reason about the map updates in one place when the brush 'paint'
method is used.

This commit also introduce the ChipsetService that knows how
to talk to the chipset API.