Add map preview generation.
This commit also attempt at fixing bugs that are only visible
on firefox. However the bugs are still there but occurs less often.

Finally, the preview is still not visible from the open-map
component. But it should be trivial to get one working. We
could also have a robust solution by doing it programmatically
so that we can re-send a request if it failed.